+++ src/tiny/superblock.rs
@@ -1,16 +1,16 @@
 use serde::{Deserialize, Serialize};
 use std::io::{Read, Write};
 
-pub const BLOCK_SIZE: u64 = 4096;
+pub const BLOCK_SIZE: usize = 4096;
 
 #[derive(Deserialize, Serialize, Debug)]
 pub struct Superblock {
-    block_size: u64,
+    block_size: usize,
     block_count: u32,
 }
 
 impl Superblock {
-    pub fn new(block_size: u64, block_count: u32) -> Superblock {
+    pub fn new(block_size: usize, block_count: u32) -> Superblock {
         Self {
             block_size,
             block_count,
